Remove '-n' from PHP execution
PHP's '-n' option prevents it from parsing INI files

This is a bit of a double-edged sword — it helps to remove differences
between environments but causes some issues, e.g. PHP will not load
extensions as they are defined in INI files

I think it's better for the basic commands to do the simplest possible
thing and allow users to customise them if they need to, e.g. to run
a language via Docker